SwipeGen/requirements.txt
adb-shell
opencv-python
Pillow
numpy
transformers>=4.30.0
torch>=2.0.0
uiautomator2